+++ R/cache-helpers.R
@@ -863,7 +863,6 @@ copySingleFile <- function(from = NULL, to = NULL, useRobocopy = TRUE,
 copyFile <- Vectorize(copySingleFile, vectorize.args = c("from", "to"))
 
 #' @rdname cacheHelper
-#' @importFrom fastdigest fastdigest
 #' @importFrom methods slotNames
 #' @importFrom digest digest
 #' @importFrom raster res crs extent
@@ -874,8 +873,9 @@ copyFile <- Vectorize(copySingleFile, vectorize.args = c("from", "to"))
   sn <- sn[!(sn %in% c("min", "max", "haveminmax", "names", "isfactor",
                        "dropped", "nlayers", "fromdisk", "inmemory", "offset", "gain"))]
   dataSlotsToDigest <- lapply(sn, function(s) slot(object@data, s))
-  dig <- fastdigest(append(list(dim(object), res(object), crs(object),
-                         extent(object)), dataSlotsToDigest)) # don't include object@data -- these are volatile
+  dig <- digest(append(list(dim(object), res(object), crs(object),
+                         extent(object)), dataSlotsToDigest),
+                algo = algo) # don't include object@data -- these are volatile
   if (nzchar(object@file@name)) {
     # if the Raster is on disk, has the first length characters;
     filename <- if (isTRUE(endsWith(basename(object@file@name), suffix = ".grd"))) {
@@ -887,7 +887,7 @@ copyFile <- Vectorize(copySingleFile, vectorize.args = c("from", "to"))
     dig2 <- .robustDigest(asPath(filename, 2), length = length, quick = quick, algo = algo)
     dig <- c(dig, dig2)
   }
-  dig <- fastdigest(dig)
+  dig <- digest(dig, algo = algo)
 }
 
 #' Recursive copying of nested environments, and other "hard to copy" objects