+++ NEWS.md
@@ -16,6 +16,7 @@ version 0.2.5.9000
 * updated `assessDataType` to consolidate `assessDataTypeGDAL` and `assessDataType` into single function (#71, @ianmseddy)
 * `cc`: new function -- a shortcut for some commonly used options for `clearCache()`
 * added experimental capacity for `prepInputs` to handle `.rar` archives, on systems with correct binaries to deal with them. Thanks to PR #86 from Tati Micheletti.
+* remove `fastdigest::fastdigest` as it is not return the identical hash across operating systems
 
 ## Bug fixes